requirements.txt
requests
pandas
PyYAML
python-dotenv
openpyxl
streamlit
plotly
